fix: move IGNORED_KEYS to constants.ts
