+++ indexer/tests/log.test.ts
@@ -1,8 +1,7 @@
 import { assertEquals } from "https://deno.land/std@0.213.0/assert/mod.ts";
-import { fromJsonRpcLog, IGNORED_KEYS, toEthLog } from "../src/types/log.ts";
-import { JsonRpcLog } from "../src/types/types.ts";
-import { bigIntToHex, Event, JsonRpcTx } from "../src/deps.ts";
-import { KAKAROT_ADDRESS } from "../src/constants.ts";
+import { fromJsonRpcLog, JsonRpcLog, toEthLog } from "./log.ts";
+import { bigIntToHex, Event, JsonRpcTx } from "../deps.ts";
+import { IGNORED_KEYS, KAKAROT_ADDRESS } from "../constants.ts";
 
 // Mock for hexToBytes
 const mockHexToBytes = (hex: string): Uint8Array => {
